@@ -11,25 +11,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.2.1",
+    "@angular/animations": "^14.2.12",
     "@angular/cdk": "^13.2.1",
-    "@angular/common": "^13.2.1",
-    "@angular/compiler": "^13.2.1",
-    "@angular/core": "^13.2.1",
-    "@angular/forms": "^13.2.1",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
     "@angular/material": "^13.2.1",
-    "@angular/platform-browser": "^13.2.1",
-    "@angular/platform-browser-dynamic": "^13.2.1",
-    "@angular/router": "^13.2.1",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "moment": "^2.29.1",
     "rxjs": "^6.5.5",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.2",
-    "@angular/cli": "^13.2.2",
-    "@angular/compiler-cli": "^13.2.1",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
     "@types/node": "^12.11.1",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.3",
@@ -44,6 +44,6 @@
     "protractor": "^7.0.0",
     "ts-node": "^8.3.0",
     "tslint": "^6.1.0",
-    "typescript": "^4.1.6"
+    "typescript": "^4.8.4"
   }
 }
